Driver and Algorithms Classes Update
I have created a driver to manage what we want to do. Sweet and simple.
I have the algorithms reading in from the .txt file right now, and creating a 2D int array for our matrix as well as a String array for our labels.

I have also created several methods to print the arrays for trouble shooting as well as one that creates a priority queue and identifies the min edge value from the 2D array.